refactor: remove duplicate ServerPermissions type
