+++ src/users/user.types.ts
@@ -1,5 +1,5 @@
 import { GroupPermissionsMap } from "../groups/group-roles/models/group-role-permission.model";
-import { ServerPermissions } from "../server-roles/models/server-role-permission.model";
+import { ServerPermissions } from "../server-roles/models/server-permissions.type";
 import { User } from "./models/user.model";
 
 export interface UserPermissions {
